added highlighting size 0.0
